chore: Update Dockerfile and docker-compose.yml
This commit updates the Dockerfile and docker-compose.yml files. The Dockerfile has been modified to use Python 3.9-slim-buster as the base image instead of Python 3.10-slim-buster. Additionally, the docker-compose.yml file has been updated to use an environment file (.env) for the BOT_TOKEN and OPENWEATHERMAP_API_KEY variables instead of directly specifying them in the file. This improves the security and flexibility of the deployment.