chore: Add additional allowed user ID check to is_owner predicate
This commit modifies the is_owner predicate in the bot.py file to include an additional check for the user's ID. The additional_allowed_user_id variable is introduced and compared with the interaction user's ID. If the user's ID matches either the server owner's ID or the additional allowed user ID, the predicate returns True. This change provides more flexibility in determining if a user is considered the owner of the server.