chore: Modify delete_all_logs_slash function to hide message by default
This commit modifies the `delete_all_logs_slash` function in `bot.py` to set the `hide_message` parameter to `True` by default. Previously, the default value was `False`. This change ensures that when the function is called without specifying the `hide_message` parameter, the message will be hidden. This modification improves the behavior and usability of the function.