chore: Add logging_channels.json and message_logs.ndjson
This commit adds the logging_channels.json file and modifies the message_logs.ndjson file. The logging_channels.json file is a new file that was created, while the message_logs.ndjson file was modified by removing multiple lines of data. These changes are necessary for implementing logging functionality in the application.