chore: Update compose.yaml to use latest image for bot service
This commit updates the `compose.yaml` file to use the latest version of the `getkarma/discord-bot-private` image for the `bot` service. The previous configuration used a local build, but this change ensures that the latest image is pulled from the repository. Additionally, the `labels` and `networks` sections have been added to the `bot` service configuration to enable watchtower auto-updates and specify the `discord_bot_network` network. These changes improve the deployment and networking capabilities of the application.