refactor: Update ADDITIONAL_ALLOWED_USER_ID in bot.py
This commit updates the `ADDITIONAL_ALLOWED_USER_ID` variable in `bot.py` to allow the bot owner and an additional user to use the owner command. The additional user's ID has been added to the variable. This change enhances the functionality of the bot by granting access to the owner command to multiple users.